+++ lepton-eda/scheme/schematic/ffi.scm
@@ -323,7 +323,9 @@
             schematic_window_show_all
             schematic_window_get_main_window
             schematic_window_set_main_window
+            schematic_window_get_pages
             schematic_window_get_tab_info_list
+            schematic_window_set_tab_info_list
             schematic_window_get_tab_notebook
             schematic_window_set_toolbar
 
@@ -342,25 +344,21 @@
             x_tabs_enabled
             x_tabs_hdr_set
             x_tabs_hdr_update
-            x_tabs_info_add
             x_tabs_info_cur
             x_tabs_info_find_by_page
-            x_tabs_info_find_by_wtab
-            x_tabs_info_rm
             x_tabs_nbook_create
-            x_tabs_nbook_page_add
-            x_tabs_nbook_page_close
-            x_tabs_page_on_reordered
-            x_tabs_tl_page_cur
-            x_tabs_tl_page_cur_set
-            x_tabs_tl_page_find
-            x_tabs_tl_pview_cur
-            x_tabs_tl_pview_cur_set
+            schematic_window_set_current_canvas
+            schematic_tab_info_new
+            schematic_tab_info_get_canvas
+            schematic_tab_info_set_canvas
+            schematic_tab_info_get_index
+            schematic_tab_info_set_index
             schematic_tab_info_get_page
             schematic_tab_info_set_page
-            schematic_tab_info_get_canvas
             schematic_tab_info_get_tab_widget
+            schematic_tab_info_set_tab_widget
             schematic_tab_info_get_window
+            schematic_tab_info_set_window
             schematic_tabs_add_canvas
             schematic_tabs_set_callback
 
@@ -620,9 +618,6 @@
             schematic_multiattrib_widget_update
 
             x_print
-
-            x_tabs_next
-            x_tabs_prev
 
             a_zoom
 
@@ -1006,7 +1001,9 @@
 (define-lff schematic_window_show_all void '(* *))
 (define-lff schematic_window_get_main_window '* '(*))
 (define-lff schematic_window_set_main_window '* '(* *))
+(define-lff schematic_window_get_pages '* '(*))
 (define-lff schematic_window_get_tab_info_list '* '(*))
+(define-lff schematic_window_set_tab_info_list void '(* *))
 (define-lff schematic_window_get_tab_notebook '* '(*))
 (define-lff schematic_window_set_toolbar void '(* *))
 
@@ -1027,26 +1024,22 @@
 (define-lff x_tabs_enabled int '())
 (define-lff x_tabs_hdr_set void '(* *))
 (define-lff x_tabs_hdr_update void '(* *))
-(define-lff x_tabs_info_add '* (list '* int '* '* '*))
 (define-lff x_tabs_info_cur '* '(*))
 (define-lff x_tabs_info_find_by_page '* '(* *))
-(define-lff x_tabs_info_find_by_wtab '* '(* *))
-(define-lff x_tabs_info_rm void '(* *))
 (define-lff x_tabs_nbook_create '* '(* *))
-(define-lff x_tabs_nbook_page_add int '(* * * *))
-(define-lff x_tabs_nbook_page_close void '(* *))
-(define-lff x_tabs_page_on_reordered void (list '* '* int '*))
 (define-lff schematic_tabs_add_canvas void '(* *))
-(define-lff x_tabs_tl_page_cur '* '(*))
-(define-lff x_tabs_tl_page_cur_set void '(* *))
-(define-lff x_tabs_tl_page_find int '(* *))
-(define-lff x_tabs_tl_pview_cur '* '(*))
-(define-lff x_tabs_tl_pview_cur_set void '(* *))
+(define-lff schematic_window_set_current_canvas void '(* *))
+(define-lff schematic_tab_info_new '* '())
+(define-lff schematic_tab_info_get_canvas '* '(*))
+(define-lff schematic_tab_info_set_canvas void '(* *))
+(define-lff schematic_tab_info_get_index int '(*))
+(define-lff schematic_tab_info_set_index void (list '* int))
 (define-lff schematic_tab_info_get_page '* '(*))
 (define-lff schematic_tab_info_set_page void '(* *))
-(define-lff schematic_tab_info_get_canvas '* '(*))
 (define-lff schematic_tab_info_get_tab_widget '* '(*))
+(define-lff schematic_tab_info_set_tab_widget void '(* *))
 (define-lff schematic_tab_info_get_window '* '(*))
+(define-lff schematic_tab_info_set_window void '(* *))
 (define-lff schematic_tabs_set_callback void '(* *))
 
 ;;; find_text_state.c
@@ -1324,8 +1317,6 @@
 (define-lff lepton_slot_update_object void '(*))
 
 ;;; x_tabs.c
-(define-lff x_tabs_next void '(*))
-(define-lff x_tabs_prev void '(*))
 
 ;;; o_undo.c
 (define-lff o_undo_savestate void (list '* '* int))