+++ lepton-eda/scheme/schematic/ffi/gtk.scm
@@ -66,8 +66,12 @@
             gtk_menu_new
             gtk_menu_shell_append
 
+            gtk_notebook_append_page
             gtk_notebook_get_n_pages
+            gtk_notebook_next_page
             gtk_notebook_page_num
+            gtk_notebook_prev_page
+            gtk_notebook_remove_page
             gtk_notebook_set_current_page
             gtk_notebook_set_tab_reorderable
 
@@ -156,8 +160,12 @@
 (define-lff gtk_menu_item_set_submenu void '(* *))
 (define-lff gtk_menu_shell_append void '(* *))
 
+(define-lff gtk_notebook_append_page int '(* * *))
 (define-lff gtk_notebook_get_n_pages int '(*))
+(define-lff gtk_notebook_next_page void '(*))
 (define-lff gtk_notebook_page_num int '(* *))
+(define-lff gtk_notebook_prev_page void '(*))
+(define-lff gtk_notebook_remove_page void (list '* int))
 (define-lff gtk_notebook_set_current_page void (list '* int))
 (define-lff gtk_notebook_set_tab_reorderable void (list '* '* int))