+++ libleptongui/include/window.h
@@ -1,7 +1,7 @@
 /* Lepton EDA Schematic Capture
  * Copyright (C) 1998-2010 Ales Hvezda
  * Copyright (C) 1998-2016 gEDA Contributors
- * Copyright (C) 2017-2024 Lepton EDA Contributors
+ * Copyright (C) 2017-2026 Lepton EDA Contributors
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -243,6 +243,9 @@ schematic_window_free (SchematicWindow *w_current);
 SchematicCanvas*
 schematic_window_get_current_canvas (SchematicWindow *w_current);
 
+void
+schematic_window_set_current_canvas (SchematicWindow* w_current,
+                                     SchematicCanvas* pview);
 GtkListStore*
 schematic_window_get_dash_length_list_store (SchematicWindow *w_current);